Reworked flow engine. Renamed some structures
