+++ src/main/kotlin/de/craftan/commands/StructureCommand.kt
@@ -5,8 +5,8 @@ import com.mojang.brigadier.arguments.IntegerArgumentType
 import com.mojang.brigadier.arguments.StringArgumentType
 import com.sk89q.worldedit.math.BlockVector3
 import de.craftan.Craftan
-import de.craftan.game.lobby.CraftanLobby
-import de.craftan.game.map.maps.DefaultMapLayout
+import de.craftan.engine.lobby.CraftanLobby
+import de.craftan.engine.map.maps.DefaultMapLayout
 import de.craftan.structures.loadStructureToClipboard
 import de.craftan.structures.placeStructure
 import net.axay.kspigot.commands.*
