+++ src/main/kotlin/de/craftan/engine/CraftanActionCard.kt
@@ -1,0 +1,3 @@
+package de.craftan.engine
+
+interface CraftanActionCard
