+++ src/main/kotlin/de/craftan/engine/CraftanGameExtension.kt
@@ -1,0 +1,11 @@
+package de.craftan.engine
+
+/**
+ * Models an extension to the base craftan game
+ */
+interface CraftanGameExtension {
+    /**
+     * The current game, this extension is adding on top of
+     */
+    val game: CraftanGame
+}
