+++ src/main/kotlin/de/craftan/engine/CraftanResource.kt
@@ -1,0 +1,7 @@
+package de.craftan.engine
+
+import org.bukkit.Material
+
+interface CraftanResource {
+    val material: Material
+}
