+++ src/main/kotlin/de/craftan/engine/CraftanStructure.kt
@@ -1,0 +1,8 @@
+package de.craftan.engine
+
+interface CraftanStructure {
+    /**
+     * Cost to build this given structure
+     */
+    val cost: Map<CraftanResource, Int>
+}
