+++ src/main/kotlin/de/craftan/engine/events/GameStartEvent.kt
@@ -1,0 +1,8 @@
+package de.craftan.engine.events
+
+import de.craftan.engine.CancelableCraftanEvent
+import de.craftan.engine.CraftanGame
+
+data class GameStartEvent(
+    override val game: CraftanGame,
+) : CancelableCraftanEvent(game)
