+++ src/main/kotlin/de/craftan/engine/events/actions/RolledDiceEvent.kt
@@ -1,0 +1,12 @@
+package de.craftan.engine.events.actions
+
+import de.craftan.engine.CraftanEvent
+import de.craftan.engine.CraftanGame
+import de.craftan.engine.CraftanPlayer
+import de.craftan.engine.map.DiceNumber
+
+data class RolledDiceEvent(
+    override val game: CraftanGame,
+    val result: DiceNumber,
+    val player: CraftanPlayer,
+) : CraftanEvent(game)
