+++ src/main/kotlin/de/craftan/engine/implementations/PreGameRound.kt
@@ -1,0 +1,13 @@
+package de.craftan.engine.implementations
+
+import de.craftan.engine.CraftanPlayer
+import de.craftan.engine.GameRound
+import de.craftan.engine.RoundFlow
+
+class PreGameRound(
+    override val player: CraftanPlayer,
+    override val flow: RoundFlow,
+) : GameRound {
+    var diceResult: Int = 0
+    override val index: Int = -1
+}
