+++ src/main/kotlin/de/craftan/engine/lobby/CraftanLobby.kt
@@ -1,0 +1,10 @@
+package de.craftan.engine.lobby
+
+import com.sk89q.worldedit.math.BlockVector3
+import com.sk89q.worldedit.world.World
+
+data class CraftanLobby(
+    val world: World,
+    val center: BlockVector3,
+    val spacing: Int,
+)
