+++ src/main/kotlin/de/craftan/engine/map/DiceNumber.kt
@@ -1,0 +1,17 @@
+package de.craftan.engine.map
+
+enum class DiceNumber(
+    val value: Int,
+) {
+    TWO(2),
+    THREE(3),
+    FOUR(4),
+    FIVE(5),
+    SIX(6),
+    SEVEN(7),
+    EIGHT(8),
+    NINE(9),
+    TEN(10),
+    ELEVEN(11),
+    TWELVE(12),
+}
