+++ src/main/kotlin/de/craftan/engine/map/GameTile.kt
@@ -1,0 +1,6 @@
+package de.craftan.engine.map
+
+data class GameTile(
+    val type: MaterialType,
+    val chance: DiceNumber,
+)
