+++ src/main/kotlin/de/craftan/engine/map/LayoutRow.kt
@@ -1,0 +1,7 @@
+package de.craftan.engine.map
+
+data class LayoutRow<T>(
+    val tiles: List<T>,
+)
+
+fun <T> layoutRow(vararg tiles: T): LayoutRow<T> = LayoutRow<T>(tiles.toList())
