+++ src/main/kotlin/de/craftan/engine/map/MaterialType.kt
@@ -1,0 +1,10 @@
+package de.craftan.engine.map
+
+enum class MaterialType {
+    WHEAT,
+    STONE,
+    SHEEP,
+    BRICK,
+    WOOD,
+    DESERT,
+}
