+++ src/main/kotlin/de/craftan/engine/map/tiles/DefaultGameTiles.kt
@@ -1,0 +1,7 @@
+package de.craftan.engine.map.tiles
+
+import de.craftan.engine.map.GameTile
+
+enum class DefaultGameTiles(
+    val tile: GameTile,
+)
