+++ src/main/kotlin/de/craftan/engine/resources/WoodResource.kt
@@ -1,0 +1,8 @@
+package de.craftan.engine.resources
+
+import de.craftan.engine.CraftanResource
+import org.bukkit.Material
+
+object WoodResource : CraftanResource {
+    override val material: Material = Material.OAK_LOG
+}
