+++ src/main/kotlin/de/craftan/engine/structures/Settlement.kt
@@ -1,0 +1,9 @@
+package de.craftan.engine.structures
+
+import de.craftan.engine.CraftanResource
+import de.craftan.engine.CraftanStructure
+import de.craftan.engine.resources.WoodResource
+
+class Settlement : CraftanStructure {
+    override val cost: Map<CraftanResource, Int> = mapOf((WoodResource to 2))
+}
