--- src/main/kotlin/de/craftan/game/CraftanActionCard.kt
@@ -1,4 +1,0 @@
-package de.craftan.game
-
-interface CraftanActionCard {
-}
