--- src/main/kotlin/de/craftan/game/CraftanResource.kt
@@ -1,7 +1,0 @@
-package de.craftan.game
-
-import org.bukkit.Material
-
-interface CraftanResource {
-    val material: Material
-}
