--- src/main/kotlin/de/craftan/game/CraftanStructure.kt
@@ -1,8 +1,0 @@
-package de.craftan.game
-
-interface CraftanStructure {
-    /**
-     * Cost to build this given structure
-     */
-    val cost: Map<CraftanResource, Int>
-}
