--- src/main/kotlin/de/craftan/game/events/GameStartEvent.kt
@@ -1,8 +1,0 @@
-package de.craftan.game.events
-
-import de.craftan.game.CancelableGameEvent
-import de.craftan.game.CraftanGame
-
-data class GameStartEvent(
-    override val game: CraftanGame,
-) : CancelableGameEvent(game)
