--- src/main/kotlin/de/craftan/game/lobby/CraftanLobby.kt
@@ -1,10 +1,0 @@
-package de.craftan.game.lobby
-
-import com.sk89q.worldedit.math.BlockVector3
-import com.sk89q.worldedit.world.World
-
-data class CraftanLobby(
-    val world: World,
-    val center: BlockVector3,
-    val spacing: Int,
-)
