--- src/main/kotlin/de/craftan/game/map/DiceNumber.kt
@@ -1,17 +1,0 @@
-package de.craftan.game.map
-
-enum class DiceNumber(
-    val value: Int,
-) {
-    TWO(2),
-    THREE(3),
-    FOUR(4),
-    FIVE(5),
-    SIX(6),
-    SEVEN(7),
-    EIGHT(8),
-    NINE(9),
-    TEN(10),
-    ELEVEN(11),
-    TWELVE(12),
-}
