--- src/main/kotlin/de/craftan/game/map/GameTile.kt
@@ -1,6 +1,0 @@
-package de.craftan.game.map
-
-data class GameTile(
-    val type: MaterialType,
-    val chance: DiceNumber,
-)
