--- src/main/kotlin/de/craftan/game/map/LayoutRow.kt
@@ -1,7 +1,0 @@
-package de.craftan.game.map
-
-data class LayoutRow<T>(
-    val tiles: List<T>,
-)
-
-fun <T> layoutRow(vararg tiles: T): LayoutRow<T> = LayoutRow<T>(tiles.toList())
