--- src/main/kotlin/de/craftan/game/map/MaterialType.kt
@@ -1,10 +1,0 @@
-package de.craftan.game.map
-
-enum class MaterialType {
-    WHEAT,
-    STONE,
-    SHEEP,
-    BRICK,
-    WOOD,
-    DESERT,
-}
