--- src/main/kotlin/de/craftan/game/map/tiles/DefaultGameTiles.kt
@@ -1,7 +1,0 @@
-package de.craftan.game.map.tiles
-
-import de.craftan.game.map.GameTile
-
-enum class DefaultGameTiles(
-    val tile: GameTile,
-)
