--- src/main/kotlin/de/craftan/game/resources/WoodResource.kt
@@ -1,8 +1,0 @@
-package de.craftan.game.resources
-
-import de.craftan.game.CraftanResource
-import org.bukkit.Material
-
-object WoodResource : CraftanResource {
-    override val material: Material = Material.OAK_LOG
-}
